requirements.txt
# Dependencies for the IEEE Quantum Week 2026 TD-TSP benchmark suite.
# Install with: pip install -r requirements.txt

# Core
numpy>=1.24.0
pandas>=2.3.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0

# Real-world TD-TSP distance matrix (Google Maps Distance Matrix API)
googlemaps>=4.10.0

# Classical solver
gurobipy>=13.0.0

# Quantum annealing (D-Wave hybrid CQM)
dwave-ocean-sdk>=6.0.0

# Gate-based quantum (AWS Braket: SV1, Rigetti)
amazon-braket-sdk>=1.50.0
boto3>=1.28.0
